// lib/FileManager/FileManager.d.ts
export type ImageOptions = {
    keepMetadata?: boolean;
    compression?: 
    /**
     * Will resize image maintaing scale ratio
     */
    {
        inside: {
            width: number;
            height: number;
        };
    } | {
        contain: {
            width: number;
        } | {
            height: number;
        };
    };
};
export type S3Config = {
    region: string;
    bucket: string;
    accessKeyId: string;
    secretAccessKey: string;
};
export type LocalConfig = {
    /**
     * example: path.join(__dirname+'/media')
     * note that this location will be relative to the compiled file location
     */
    localFolderPath: string;
    /**
     * Minimum amount of free bytes available to allow saving files
     * Defaults to 100MB
     */
    minFreeBytes?: number;
};
export type UploadItem = {
    name: string;
    content_type: string;
    data: Buffer;
    extension: string;
};
export type UploadedItem = {
    /**
     * Url that is passed to client
     */
    url: string;
    filePath?: string;
    etag: string;
    /**
     * S3 url of the resource
     */
    s3_url?: string;
    /**
     * Total uploaded file size in bytes
     */
    content_length: number;
};
export declare class FileManager {
    static testCredentials: (accessKeyId: string, secretAccessKey: string) => Promise<import("aws-sdk/lib/request").PromiseResult<AWS.STS.GetCallerIdentityResponse, AWS.AWSError>>;
    s3Client?: S3;
    config: S3Config | LocalConfig;
    imageOptions?: ImageOptions;
    prostgles?: Prostgles;
    get dbo(): DBHandlerServer;
    get db(): DB;
    tableName?: string;
    fileRoute?: string;
    get fileRouteExpress(): string;
    private checkInterval?;
    constructor(config: FileManager["config"], imageOptions?: ImageOptions);
    getFileStream(name: string): Promise<stream.Readable>;
    deleteFile(name: string): Promise<true | import("aws-sdk/lib/request").PromiseResult<S3.DeleteObjectOutput, AWS.AWSError>>;
    parseFile: any;
    getLocalFileUrl: (name: string) => string;
    checkFreeSpace: (folderPath: string, fileSize?: number) => Promise<void>;
    uploadStream: any;
    upload: any;
    uploadAsMedia: (params: {
        item: UploadItem;
        allowedExtensions?: Array<ALLOWED_EXTENSION>;
        dissallowedExtensions?: Array<ALLOWED_EXTENSION>;
        imageOptions?: ImageOptions;
    }) => Promise<UploadedItem>;
    getFileS3URL(fileName: string, expiresInSeconds?: number): Promise<string>;
    parseSQLIdentifier: (name: string) => Promise<string>;
    getColInfo: (args: {
        tableName: string;
        colName: string;
    }) => ValidatedColumnInfo["file"] | undefined;
    init: any;
    destroy: () => void;
}